[WIP] Refactors i18n inclusion so that client and server configuration do not collide.
