+++ frontend/pages/+Layout.tsx
@@ -1,18 +1,6 @@
-import { type FC, type ReactNode, use } from "react";
-import { I18nextProvider } from "react-i18next";
-import i18n from "./i18n";
+import { type FC, type ReactNode } from "react";
 import "./index.scss";
 
-const translate = i18n.loadNamespaces("translation");
-
-const Layout: FC<{ children: ReactNode }> = ({ children }) => {
-  use(translate);
-  return (
-    <div className="layout">
-      <I18nextProvider i18n={i18n} defaultNS={"translation"}>
-        {children}
-      </I18nextProvider>
-    </div>
-  );
-};
-export default Layout;
+export const Layout: FC<{ children: ReactNode }> = ({ children }) => (
+  <div className="layout">{children}</div>
+);
